requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0

# ─── Twitter ─────────────────────────────────────────────────────────────────
tweepy>=4.14.0

# ─── Telegram ────────────────────────────────────────────────────────────────
# Uses requests (already above) – no extra package needed

# ─── Web Dashboard ───────────────────────────────────────────────────────────
flask>=3.0.0
